fixes javascipt bug in parsing frets
--there was a bug when the javascript was
interpreting frets that are with two digits
size. In that case a separator - is used in
order to separate the fret for each string
but the javascript didnt take account of that
and an error was appearing. This commit fixes
that issue, now frets longer than 9th fill
the boxes correctly